Update sheetFiller.py
Added a few features:
1. if you write a typo/non-existant contract name - the cell will be marked in red (and the script will keep running without an error)
once you fix it the cell changes back to black borders.
could not make it modular to any cell so its fixed on the right cell.

2. added system to check if more than one contract fits the written name (for example the "V2" contracts we have this saturday).
Once it happens - it checks if any of them fits the exact name (so if you write the exact contract name it will choose it, even in another contract contains the name within its own (see "sand spider" and "sand spider V2").
If there is no exact match - it will take the newest.

3. Log file now gets a custom log that fits the situation (found X contracts, chose the exact/the latest).

4. Fixed error if a contract has no responses when its entered.

Also made sure all of that has minimal impact on the DB and the Sheet, so it doesn't spam them.